Fix compiler warnings.

+++ jEditLauncher/trunk/src/jecfg.cpp
@@ -23,7 +23,7 @@ int WINAPI _tWinMain(HINSTANCE hInstance, HINSTANCE /*hPrevInstance*/, LPTSTR lp
     // BLOCK: Run application
     {
         CMainDlg dlgMain;
-        nRet = dlgMain.DoModal();
+        nRet = (int)dlgMain.DoModal();
     }
 
     _Module.Term();
